* utilizing world's array of sprites, the world stores all the sprite information (fileName, frameWidth, frameHeight, animSpeed, rows), but each entity has to create a new sprite for itself, using 1 of the known sprites

+++ Team BMP/Minion.h
@@ -106,9 +106,9 @@ public:
 		for(int i = 0; i < NUM_STATS; i++)
 		{
 			if(m_eType == MINION)//scale the minions slightly weeker then player but close
-				m_stats[i] = a_player->getStatNumber(i) * SCALE_MIN;
+				m_stats[i] = (int)(a_player->getStatNumber(i) * SCALE_MIN);
 			else if( m_eType == BOSS)
-				m_stats[i] = a_player->getStatNumber(i) * SCALE_BOSS;
+				m_stats[i] = (int)(a_player->getStatNumber(i) * SCALE_BOSS);
 		}
 		m_stats[HEALTH_CURRENT] = m_stats[HEALTH_MAX];//incase the players health is low this will set the minion/boss to full
 	}
